+++ CHANGELOG.md
@@ -43,6 +43,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The Voluntary project creation journey now has two different "endings"
   depending on whether the project is being handed over to Regional Caseworker
   services or not
+- Removed the Caseworker and Team Leader fields from the internal contact tab
 
 ### Added